Fix stack tests
