+++ test/runtests.jl
@@ -5,31 +5,33 @@ include("resources/TestRequest.jl")
 
 @testset "HTTP" begin
     for f in [
-              # "ascii.jl",
-              # "chunking.jl",
-              # "utils.jl",
-              # "client.jl",
-              # "multipart.jl",
-              # "parsemultipart.jl",
-              # "sniff.jl",
-              # "cookies.jl",
-              # "parser.jl",
-              # "loopback.jl",
-              # "websockets.jl",
-              # "messages.jl",
-              # "handlers.jl",
-              # "server.jl",
-              # "async.jl",
-              # "aws4.jl",
-              "insert_layers.jl",
-              # "mwe.jl",
+              "ascii.jl",
+              "chunking.jl",
+              "utils.jl",
+              "client.jl",
+              "multipart.jl",
+              "parsemultipart.jl",
+              "sniff.jl",
+              "cookies.jl",
+              "parser.jl",
+              "loopback.jl",
+              "websockets.jl",
+              "messages.jl",
+              "handlers.jl",
+              "server.jl",
+              "async.jl",
+              "aws4.jl",
+              "stack.jl",
+              "mwe.jl",
              ]
         file = joinpath(dir, f)
         println("Running $file tests...")
         if isfile(file)
-            include(file)
+            # include(file)
         else
             @show readdir(dirname(file))
         end
     end
+    # TODO: Set file back in the loop
+    include("stack.jl")
 end
